fix: allow DM user to search the content
DM has a several user roles that not by default can search the content on their channel except the owner. The user roles are admin and editor. Need a special treatment to do that. First, get the channel that belongs to the user. This feature is hidden by default and not published on the document.

Need adjust something on the back end and the front that call the channel to search the content. Files affected are `connect_page.php`, `VideoComponent.js`, and `PlaylistComponent.js`.