fix: update third party packages against PHPCS
This plugin is depends on 2 third party packages. The dependency is not safe to use, the PHPCS throw errors. The packages are below

- PHP-DM-SDK
- PluginUpdateChecker

Since the `PHP-DM-SDK` is not use in any of the plugin, we can remove it. The `PluginUpdateChecker` need to update to our own version and send the PR to the owner to make it available in public. These time, we're pointing to our github repo until the PR merged.

The `autoload` for external packages (using git repo services like github or gitlab) needs extra configuration to make it works. It's not by default auto loaded. The references below.

References:
1. https://stackoverflow.com/questions/19566102/github-repository-autoloading-with-composer-issue
2. https://getcomposer.org/doc/04-schema.md#autoload